GYG PROJE/survey_app_insecurty/requirements.txt
python-dotenv
Flask
Flask-Login
Flask-WTF
Flask-MySQLdb
